requirements.txt
numpy
pandas
matplotlib
python-dateutil
requests
google-cloud-storage
psutil